pyproject.toml
[project]
name = "projeto-etl-contadoria"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gdown>=5.2.1",
    "google-api-python-client>=2.193.0",
    "google-auth>=2.49.1",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "gspread>=6.2.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pydrive>=1.3.1",
    "pydrive2>=1.21.3",
    "pyotp>=2.9.0",
    "python-dotenv>=1.2.2",
    "selenium>=4.41.0",
    "supabase>=2.28.3",
    "webdriver-manager>=4.0.2",
    "apache-airflow>=2.10.5",
    "psycopg2-binary>=2.9.9",
]

[tool.uv.pip]
system = true
